@@ -18,7 +18,7 @@ import org.appcelerator.titanium.view.TiDrawableReference;
 import ti.modules.titanium.android.AndroidModule;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.support.v4.app.NotificationCompat.BigPictureStyle;
+import androidx.core.app.NotificationCompat.BigPictureStyle;
 
 @Kroll.proxy(creatableInModule = AndroidModule.class, propertyAccessors = { TiC.PROPERTY_DECODE_RETRIES })
 public class BigPictureStyleProxy extends StyleProxy
