@@ -38,7 +38,7 @@ import android.hardware.Camera.Size;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.ActionBar;
+import androidx.appcompat.app.ActionBar;
 import android.view.Gravity;
 import android.view.KeyEvent;
 import android.view.Surface;
