@@ -19,9 +19,8 @@ import ti.modules.titanium.ui.widget.webview.TiUIWebView;
 import android.app.Activity;
 import android.content.Intent;
 import android.graphics.PixelFormat;
-import android.text.util.Linkify;
 import android.view.Gravity;
-import android.support.v4.view.GravityCompat;
+import androidx.core.view.GravityCompat;
 import android.view.WindowManager;
 import android.webkit.WebSettings;
 // clang-format off
