@@ -9,15 +9,12 @@ package ti.modules.titanium.ui.android;
 import ti.modules.titanium.ui.widget.TiUIDrawerLayout;
 
 import org.appcelerator.kroll.annotations.Kroll;
-import org.appcelerator.kroll.common.Log;
-import org.appcelerator.kroll.KrollProxy;
 import org.appcelerator.titanium.proxy.TiViewProxy;
 import org.appcelerator.titanium.TiC;
-import org.appcelerator.titanium.util.TiConvert;
 import org.appcelerator.titanium.view.TiUIView;
 
 import android.app.Activity;
-import android.support.v4.widget.DrawerLayout;
+import androidx.drawerlayout.widget.DrawerLayout;
 
 @Kroll.proxy(creatableInModule = AndroidModule.class)
 public class DrawerLayoutProxy extends TiViewProxy
