@@ -9,8 +9,8 @@ package org.appcelerator.titanium.view;
 
 import android.content.res.Configuration;
 import android.content.res.TypedArray;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.View;
 import android.view.ViewGroup;
 import org.appcelerator.kroll.common.Log;
