@@ -62,7 +62,7 @@ class AndroidManifest {
 
 	/**
 	 * Creates an XML string of the manifest contents.
-	 * @returns {String} Returns an XML string of the Android manfiest contents. Will never return null.
+	 * @returns {String} Returns an XML string of the Android manifest contents. Will never return null.
 	 */
 	toString() {
 		let text = '<?xml version="1.0" encoding="utf-8"?>' + os.EOL;
@@ -208,6 +208,27 @@ class AndroidManifest {
 
 		// Assign package name to <manifest/> element.
 		manifestElement.setAttribute('package', name);
+	}
+
+	/**
+	 * Fetches the "versionCode" and "versionName" attributes from the <manifest/> element.
+	 * @return {{versionCode: {String}, versionName: {String}}}
+	 * Returns a dictionary providing the "versionCode" and "versionName" attributes read from <manifest/> element.
+	 *
+	 * Returns null if no XML content has been loaded;
+	 */
+	getAppVersionInfo() {
+		let versionInfo = null;
+		if (!this.isEmpty()) {
+			const element = this._xmlDomDocument.documentElement;
+			if (element.tagName === 'manifest') {
+				versionInfo = {
+					versionCode: element.getAttribute('android:versionCode'),
+					versionName: element.getAttribute('android:versionName')
+				};
+			}
+		}
+		return versionInfo;
 	}
 
 	/**
@@ -598,20 +619,30 @@ class AndroidManifest {
 			throw new Error('1st argument must be of type string.');
 		}
 
-		// Inject commonly used Android XML namespaces into the <manifest/> element if missing.
-		// Note: Needed for backward compatibility since older versions of Titanium would inject them too.
+		// Process the XML string before parsing it.
 		const MANIFEST_ELEMENT_PREFIX = '<manifest';
 		const manifestIndex = text.indexOf(MANIFEST_ELEMENT_PREFIX);
 		if (manifestIndex >= 0) {
+			// Inject commonly used Android XML namespaces into the <manifest/> element if missing.
+			// Needed for backward compatibility since older versions of Titanium would inject them too.
 			const insertIndex = manifestIndex + MANIFEST_ELEMENT_PREFIX.length;
-			if (text.indexOf('xmlns:tools') < 0) {
+			let attributeIndex = text.indexOf('xmlns:tools', manifestIndex);
+			if ((attributeIndex < 0) || (attributeIndex > text.indexOf('>', manifestIndex))) {
 				const attributeString = ' xmlns:tools="http://schemas.android.com/tools"';
 				text = text.substring(0, insertIndex) + attributeString + text.substring(insertIndex);
 			}
-			if (text.indexOf('xmlns:android') < 0) {
+			attributeIndex = text.indexOf('xmlns:android', manifestIndex);
+			if ((attributeIndex < 0) || (attributeIndex > text.indexOf('>', manifestIndex))) {
 				const attributeString = ' xmlns:android="http://schemas.android.com/apk/res/android"';
 				text = text.substring(0, insertIndex) + attributeString + text.substring(insertIndex);
 			}
+
+			// Remove "xmlns:android" namespace attributes from all child elements. Only supported in <manifest/>.
+			// Note: CLI used to wrongly inject these after updating "tiapp.xml" if missing from root element.
+			const manifestEndIndex = text.indexOf('>', manifestIndex);
+			text = text.replace(/xmlns:android=".*?"/g, (match, p1) => {
+				return (p1 <= manifestEndIndex) ? match : '';
+			});
 		}
 
 		// Parse given XML string and wrap it in a new AndroidManifest instance.
