@@ -25,6 +25,19 @@ const url = require('url');
  */
 const isWindows = (process.platform === 'win32');
 
+/**
+ * Array of dictionary objects describing what is to be written to a Gradle/Java style properties file.
+ * Order matters. First array entry is written to file first. Last array entry is written last.
+ *
+ * Null or empty entries will write empty lines to file.
+ *
+ * An entry with a "comment" property will write a #-prefixed string to file.
+ *
+ * An entry having "key" and "value" properties will be written to file as "myKey=myValue".
+ * You can set a "key" without a "value" which will write "myKey=" to file.
+ * @typedef {[{ comment: {String}, key: {String}, value: {String} }]} GradlePropertiesFileEntries
+ */
+
 /** Class used to install, configure, and run gradle in a root project directory. */
 class GradleWrapper {
 	/**
@@ -218,7 +231,15 @@ class GradleWrapper {
 		await fs.copy(templateDirPath, destinationDirPath);
 		await fs.chmod(path.join(destinationDirPath, 'gradlew'), 0o755);
 		await fs.chmod(path.join(destinationDirPath, 'gradlew.bat'), 0o755);
+	}
 
+	/**
+	 * Creates an array of property entries to be written to a "gradle.properties" file.
+	 * Provides default property key/value entries such as network proxy settings.
+	 * @returns {Promise<GradlePropertiesFileEntries>}
+	 * Returns an array of property file entries to be written to a "gradle.properties" file.
+	 */
+	async fetchDefaultGradleProperties() {
 		// Fetch proxy server information, if configured.
 		let proxyUrl = null;
 		await new Promise((resolve) => {
@@ -238,7 +259,7 @@ class GradleWrapper {
 			});
 		});
 
-		// Write a "gradle.properties" file.
+		// Create an array of Java property entries and return it.
 		const properties = [ { comment: '# This file was generated by Titanium\'s build tools.' } ];
 		if (proxyUrl) {
 			if (proxyUrl.hostname) {
@@ -259,7 +280,27 @@ class GradleWrapper {
 				}
 			}
 		}
-		await writeJavaPropertiesFile(path.join(destinationDirPath, 'gradle.properties'), properties);
+		return properties;
+	}
+
+	/**
+	 * Creates a "gradle.properties" file at the gradle project's root location,
+	 * configuring it with default properties such as network proxy settings.
+	 */
+	async writeDefaultGradlePropertiesFile() {
+		const properties = await this.fetchDefaultGradleProperties();
+		await this.writeGradlePropertiesFile(properties);
+	}
+
+	/**
+	 * Creates a "gradle.properties" file at the gradle project's root location,
+	 * writing the given properties key/value entries to it.
+	 * @param {GradlePropertiesFileEntries} [properties]
+	 * Array of properties to be written to file. Can be null/undefined, in which case an empty file will be created.
+	 */
+	async writeGradlePropertiesFile(properties) {
+		const filePath = path.join(this._gradlewDirPath, 'gradle.properties');
+		await writeJavaPropertiesFile(filePath, properties);
 	}
 
 	/**
@@ -299,17 +340,9 @@ function isNonEmptyString(value) {
  * If referenced file already exists, then it will be overwritten.
  * @param {String} filePath
  * Path to the file to be written. Cannot be null/undefined or else an exception will be thrown.
- * @param {[{ comment: {String}, key: {String}, value: {String} }]} properties
- * Array of property dictionary objects describing what to be written to file.
- * Order matters. First array entry is written to file first. Last array entry is written last.
- * Given array can be null/undefined, in which case an empty file will be created.
- *
- * Null or empty entries will write empty lines to file.
- *
- * An entry with a "comment" property will write a #-prefixed string to file.
- *
- * An entry having "key" and "value" properties will be written to file as "myKey=myValue".
- * You can set a "key" without a "value" which will write "myKey=" to file.
+ * @param {GradlePropertiesFileEntries} [properties]
+ * An array of properties to be written to file.
+ * Can be null/undefined, in which case an empty file will be created.
  */
 async function writeJavaPropertiesFile(filePath, properties) {
 	// Validate arguments.
@@ -353,9 +386,9 @@ async function writeJavaPropertiesFile(filePath, properties) {
 	}
 
 	// Turn the given property entries into an array of text lines to be written to file later.
-	let fileLines = [];
+	const fileLines = [];
 	if (properties) {
-		for (let nextProperty of properties) {
+		for (const nextProperty of properties) {
 			// Handle the next property row entry.
 			let hasAddedLine = false;
 			if (typeof nextProperty === 'object') {
@@ -364,7 +397,7 @@ async function writeJavaPropertiesFile(filePath, properties) {
 					let comment = nextProperty.comment;
 					comment = comment.replace(/[\r\n]/g, '');
 					if (!comment.startsWith('#') && !comment.startsWith('!')) {
-						comment = '#' + comment;
+						comment = '# ' + comment;
 					}
 					fileLines.push(comment);
 					hasAddedLine = true;
