@@ -13,8 +13,8 @@ const util = require('util');
 const exec = util.promisify(require('child_process').exec); // eslint-disable-line security/detect-child-process
 const fs = require('fs-extra');
 const glob = util.promisify(require('glob'));
-const moment = require('moment');
 const path = require('path');
+const timestamp = require('../../build/lib/utils').timestamp;
 
 // Determine if we're running on a Windows machine.
 const isWindows = (process.platform === 'win32');
@@ -81,13 +81,18 @@ async function generateBuildProperties() {
 		buildGitHash = 'HEAD';
 	}
 
+	let buildTimestamp = process.env.TI_SDK_BUILD_TIMESTAMP;
+	if (!buildTimestamp) {
+		buildTimestamp = timestamp();
+	}
+
 	// Create the "build.properties" content.
 	const fileContentString
 		= '#Generated by Titanium\n'
 		+ '\n'
 		+ 'build.version=' + escapeValue(buildVersion) + '\n'
 		+ 'build.githash=' + escapeValue(buildGitHash) + '\n'
-		+ 'build.timestamp=' + escapeValue(moment().format('YYYY/MM/DD HH:mm')) + '\n';
+		+ 'build.timestamp=' + escapeValue(buildTimestamp) + '\n';
 
 	// Create the "build.properties" file under the "assets" directory.
 	const directoryPath = path.join(__dirname, 'assets', 'Resources', 'ti.internal');
@@ -106,9 +111,21 @@ async function generateSourceCode() {
 		runtimeV8DirPath, 'src/native/KrollNativeBindings.gperf',
 		path.join(runtimeV8DirPath, 'generated/KrollNativeBindings.h'));
 
+	// Fetch the "python" command line tool to use. Prefer "python2" if available.
+	// TODO: Replace our usage of python to make it easier to build on Windows.
+	let pythonCommand = 'python';
+	if (!isWindows) {
+		try {
+			await exec('type python2 > /dev/null');
+			pythonCommand = 'python2';
+		} catch (ex) {
+			// The "python2" command line tool does not exist on this system.
+		}
+	}
+
 	// Create "bootstrap.js" which generates create*() functions for all Titanium internal modules.
 	// TODO: Replace our usage of python to make it easier to build on Windows.
-	await exec('python ' + quotePath(path.join(runtimeV8DirPath, 'tools', 'genBootstrap.py')) + ' --runtime=v8');
+	await exec(pythonCommand + ' ' + quotePath(path.join(runtimeV8DirPath, 'tools', 'genBootstrap.py')) + ' --runtime=v8');
 
 	// Generate a "KrollGeneratedBindings.h" file with perfect hashes via gperf tool.
 	// Note: 2nd argument is inserted into file as-is. Use relative path since absolute may contain user name.
@@ -138,7 +155,7 @@ async function generateSourceCode() {
 	// Generate a "KrollJS.h" file containing all of the above fetched JS files.
 	// TODO: Replace our usage of python to make it easier to build on Windows.
 	const commandLineArgs = [
-		'python',
+		pythonCommand,
 		quotePath(path.join(runtimeV8DirPath, 'tools', 'js2c.py')),
 		quotePath(path.join(runtimeV8DirPath, 'generated', 'KrollJS.h')),
 		quotePath(path.join(runtimeV8DirPath, 'generated', 'bootstrap.js'))
