@@ -16,6 +16,18 @@
 
 @implementation TiUIWebViewProxy
 
+static NSArray *webViewKeySequence;
+
+#pragma mark Internal
+
+- (NSArray *)keySequence
+{
+  if (webViewKeySequence == nil) {
+    webViewKeySequence = [[NSArray arrayWithObjects:@"assetsDirectory", @"url", nil] retain];
+  }
+  return webViewKeySequence;
+}
+
 - (id)_initWithPageContext:(id<TiEvaluator>)context
 {
   if (self = [super _initWithPageContext:context]) {
@@ -385,7 +397,11 @@
 
 - (void)reload:(id)unused
 {
-  [[[self webView] webView] reload];
+  TiThreadPerformOnMainThread(
+      ^{
+        [[self webView] reload];
+      },
+      NO);
 }
 
 - (void)repaint:(id)unused
